+++ config/urls.py
@@ -103,6 +103,10 @@ urlpatterns = [
                       "plugins/",
                       include("topobank.plugins.urls", namespace="plugins"),
                   ),
+                  path(
+                      "watchman/",
+                      include(("watchman.urls", "watchman"), namespace="watchman")
+                  ),
             ] + static(
                 settings.MEDIA_URL, document_root=settings.MEDIA_ROOT
             )