+++ topobank/taskapp/utils.py
@@ -1,6 +1,7 @@
 import importlib
 
 from django.db import transaction
+from watchman.decorators import check as watchman_check
 
 from topobank.analysis.models import Dependency, Version
 
@@ -58,3 +59,24 @@ def get_package_version_instance(pkg_name, version_expr):
     version, created = Version.objects.get_or_create(dependency=dep, major=major, minor=minor, micro=micro)
 
     return version
+
+
+def celery_worker_check():
+    return {
+        'celery': _celery_worker_check(),
+    }
+
+@watchman_check
+def _celery_worker_check():
+    """Used with watchman in order to check whether celery workers are available."""
+    # See https://github.com/mwarkentin/django-watchman/issues/8
+    from .celeryapp import app
+    MIN_NUM_WORKERS_EXPECTED = 1
+    d = app.control.broadcast('ping', reply=True, timeout=0.5, limit=MIN_NUM_WORKERS_EXPECTED)
+    return {
+        'num_workers_available': len(d),
+        'min_num_workers_expected': MIN_NUM_WORKERS_EXPECTED,
+        'ok': len(d) >= MIN_NUM_WORKERS_EXPECTED,
+    }
+
+