+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - MAINT: Upgrade of SurfaceTopograpy to 1.3.0
 - MAINT: Renamed "Sign in via ORCID" to "Sign in"
 - MAINT: Default AWS access information to None
+- MAINT: Added django-watchman, added celery check
 
 ## 0.92.0 (2022-12-14)